% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_settingvalue.R
\name{auto_settingvalue}
\alias{auto_settingvalue}
\title{Automatically Set glmnet Setting Parameters}
\usage{
auto_settingvalue(
  X,
  size = 1000,
  T_hope = 20,
  seed = 1,
  message = TRUE,
  line = TRUE
)
}
\arguments{
\item{X}{Numeric vector of input data.}

\item{size}{Integer, number of samples to generate for optimization.}

\item{T_hope}{Numeric, desired threshold time for computation.}

\item{message}{Logical, whether to display result messages.}
}
\value{
A list containing optimal \code{nlambda}, \code{thresh}, pareto front, and pareto data.
}
\description{
This function uses a neural network model to find optimal \code{nlambda} and \code{thresh} parameters based on input data.
}
